requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
aiofiles==24.1.0
    # via gradio
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   gradio
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   nltk
    #   typer
    #   uvicorn
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via fastmcp
fastapi==0.115.12
    # via
    #   mcp-agent-basic (pyproject.toml)
    #   gradio
fastmcp==2.5.2
    # via mcp-agent-basic (pyproject.toml)
ffmpy==0.5.0
    # via gradio
filelock==3.18.0
    # via huggingface-hub
fsspec==2025.5.1
    # via
    #   gradio-client
    #   huggingface-hub
gradio==5.32.0
    # via mcp-agent-basic (pyproject.toml)
gradio-client==1.10.2
    # via gradio
groovy==0.1.2
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   mcp-agent-basic (pyproject.toml)
    #   fastmcp
    #   gradio
    #   gradio-client
    #   mcp
    #   openai
    #   safehttpx
httpx-sse==0.4.0
    # via mcp
huggingface-hub==0.32.3
    # via
    #   gradio
    #   gradio-client
    #   smolagents
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   gradio
    #   smolagents
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via nltk
jsonref==1.1.0
    # via mcpadapt
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   gradio
    #   jinja2
mcp==1.9.0
    # via
    #   mcp-agent-basic (pyproject.toml)
    #   fastmcp
    #   gradio
    #   mcpadapt
    #   smolagents
mcpadapt==0.1.9
    # via smolagents
mdurl==0.1.2
    # via markdown-it-py
nltk==3.9.1
    # via textblob
numpy==2.2.6
    # via
    #   gradio
    #   pandas
openai==1.82.1
    # via smolagents
openapi-pydantic==0.5.1
    # via fastmcp
orjson==3.10.18
    # via gradio
packaging==25.0
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
pandas==2.2.3
    # via gradio
pillow==11.2.1
    # via
    #   gradio
    #   smolagents
pydantic==2.11.5
    # via
    #   fastapi
    #   gradio
    #   mcp
    #   mcpadapt
    #   openai
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via mcp
pydub==0.25.1
    # via gradio
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   fastmcp
    #   mcpadapt
    #   pydantic-settings
    #   smolagents
python-multipart==0.0.20
    # via
    #   gradio
    #   mcp
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   gradio
    #   huggingface-hub
regex==2024.11.6
    # via nltk
requests==2.32.3
    # via
    #   huggingface-hub
    #   smolagents
rich==14.0.0
    # via
    #   fastmcp
    #   smolagents
    #   typer
ruff==0.11.12
    # via gradio
safehttpx==0.1.6
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smolagents==1.17.0
    # via mcp-agent-basic (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sse-starlette==2.3.6
    # via mcp
starlette==0.46.2
    # via
    #   fastapi
    #   gradio
    #   mcp
textblob==0.19.0
    # via mcp-agent-basic (pyproject.toml)
tomlkit==0.13.2
    # via gradio
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   nltk
    #   openai
typer==0.16.0
    # via
    #   fastmcp
    #   gradio
typing-extensions==4.13.2
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
uvicorn==0.34.3
    # via
    #   gradio
    #   mcp
websockets==15.0.1
    # via
    #   fastmcp
    #   gradio-client
